Correção do tipo de dado da aliquota

+++ src/br/com/rest/model/TributoICMS.java
@@ -38,7 +38,7 @@ public class TributoICMS {
 		return aliquota;
 	}
 	
-	public void setAliquota(float aliquota) {
+	public void setAliquota(double aliquota) {
 		this.aliquota = aliquota;
 	}
